+++ src/lxc-util.c
@@ -413,16 +413,16 @@ static int lxcutil_set_config_static_device(struct lxc_container *plxc, containe
 		buflen = (ssize_t)sizeof(buf) - slen - 1;
 		if (develem->optional == 1) {
 			(void) strncpy(&buf[slen], ",optional", buflen);
-			slen = slen + sizeof(",optional") - 1u;
+			slen = slen + (ssize_t)sizeof(",optional") - 1;
 		}
 
-		buflen = sizeof(buf) - slen - 1;
+		buflen = (ssize_t)sizeof(buf) - slen - 1;
 		if (develem->type == DEVICE_TYPE_DEVNODE) {
 			(void) strncpy(&buf[slen], ",create=file", buflen);
-			slen = slen + sizeof(",create=file") - 1u;
+			slen = slen + (ssize_t)sizeof(",create=file") - 1;
 		} if (develem->type == DEVICE_TYPE_DEVDIR) {
 			(void) strncpy(&buf[slen], ",create=dir", buflen);
-			slen = slen + sizeof(",create=dir") - 1u;
+			slen = slen + (ssize_t)sizeof(",create=dir") - 1;
 		}
 
 		bret = plxc->set_config_item(plxc, "lxc.mount.entry", buf);
